+++ changelog.md
@@ -5,12 +5,17 @@
 ### Added
 
 - Added default target language setting for each german language
+- Added some new hooks
 
 ### Optimized
 
 - New default target language for new installations in any german languages: plain language
 -> reason: there are clearer guidelines for this in germany
 - Progressbar on setup now jumps to green
+
+### Fixed
+
+- Assigned page template got lost during creating a new simplification object
 
 ## [2.3.2] - 06.01.2025